Add files via upload
unordered list ( <ul> ) is used to create a list of items in no particular order i.e. the order of items is not relevant. By default, the items in this list will be marked with bullets. Whereas, an ordered list ( <ol> ) is used to create a list of items in a specific order.

Ordered lists :
Used to display items in a specific order, with each item marked with numbers by default. For example, a ranked list of movies would be a good use for an ordered list.
Unordered lists :
Used to display items in no particular order, with each item marked with bullets by default. For example, a list of ingredients to buy for baking a cake would be a good use for an unordered list.